Moved stand alone tutorials into examples/tutorials directory
