+++ examples/tutorials/stereo_agent.py
