+++ tests/test_examples.py
@@ -5,9 +5,9 @@ import subprocess
 
 import pytest
 
-import examples.new_actions
-import examples.stereo_agent
 import examples.tutorials.lighting_tutorial
+import examples.tutorials.new_actions
+import examples.tutorials.stereo_agent
 
 
 @pytest.mark.gfxtest
@@ -19,8 +19,8 @@ import examples.tutorials.lighting_tutorial
 @pytest.mark.parametrize(
     "example_module,args",
     [
-        (examples.stereo_agent, (False,)),
-        (examples.new_actions, ()),
+        (examples.tutorials.stereo_agent, (False,)),
+        (examples.tutorials.new_actions, ()),
         (examples.tutorials.lighting_tutorial, (False,)),
     ],
 )
